+++ src/core/src/sketch/minhash.rs
@@ -12,24 +12,24 @@ use serde::ser::{SerializeStruct, Serializer};
 use serde::{Deserialize, Serialize};
 use typed_builder::TypedBuilder;
 
-use crate::_hash_murmur;
 use crate::encodings::HashFunctions;
 use crate::signature::SigsTrait;
 use crate::sketch::hyperloglog::HyperLogLog;
 use crate::Error;
+use crate::{ScaledType, _hash_murmur};
 
-pub fn max_hash_for_scaled(scaled: u64) -> u64 {
+pub fn max_hash_for_scaled(scaled: ScaledType) -> u64 {
     match scaled {
-        0 => 0,
+        0 => 0, // scaled == 0 indicates this is a num minhash
         1 => u64::MAX,
         _ => (u64::MAX as f64 / scaled as f64) as u64,
     }
 }
 
-pub fn scaled_for_max_hash(max_hash: u64) -> u64 {
+pub fn scaled_for_max_hash(max_hash: u64) -> ScaledType {
     match max_hash {
-        0 => 0,
-        _ => (u64::MAX as f64 / max_hash as f64) as u64,
+        0 => 0, // scaled == 0 indicates this is a num minhash
+        _ => (u64::MAX as f64 / max_hash as f64) as ScaledType,
     }
 }
 
@@ -185,7 +185,7 @@ impl<'de> Deserialize<'de> for KmerMinHash {
 
 impl KmerMinHash {
     pub fn new(
-        scaled: u64,
+        scaled: ScaledType,
         ksize: u32,
         hash_function: HashFunctions,
         seed: u64,
@@ -230,7 +230,7 @@ impl KmerMinHash {
         self.max_hash
     }
 
-    pub fn scaled(&self) -> u64 {
+    pub fn scaled(&self) -> ScaledType {
         scaled_for_max_hash(self.max_hash)
     }
 
@@ -537,13 +537,13 @@ impl KmerMinHash {
     }
 
     pub fn count_common(&self, other: &KmerMinHash, downsample: bool) -> Result<u64, Error> {
-        if downsample && self.max_hash != other.max_hash {
-            let (first, second) = if self.max_hash < other.max_hash {
+        if downsample && self.scaled() != other.scaled() {
+            let (first, second) = if self.scaled() > other.scaled() {
                 (self, other)
             } else {
                 (other, self)
             };
-            let downsampled_mh = second.downsample_max_hash(first.max_hash)?;
+            let downsampled_mh = second.clone().downsample_scaled(first.scaled())?;
             first.count_common(&downsampled_mh, false)
         } else {
             self.check_compatible(other)?;
@@ -685,13 +685,14 @@ impl KmerMinHash {
         ignore_abundance: bool,
         downsample: bool,
     ) -> Result<f64, Error> {
-        if downsample && self.max_hash != other.max_hash {
-            let (first, second) = if self.max_hash < other.max_hash {
+        if downsample && self.scaled() != other.scaled() {
+            // downsample to larger of two scaled
+            let (first, second) = if self.scaled() > other.scaled() {
                 (self, other)
             } else {
                 (other, self)
             };
-            let downsampled_mh = second.downsample_max_hash(first.max_hash)?;
+            let downsampled_mh = second.clone().downsample_scaled(first.scaled())?;
             first.similarity(&downsampled_mh, ignore_abundance, false)
         } else if ignore_abundance || self.abunds.is_none() || other.abunds.is_none() {
             self.jaccard(other)
@@ -721,23 +722,14 @@ impl KmerMinHash {
     }
 
     // create a downsampled copy of self
-    pub fn downsample_max_hash(&self, max_hash: u64) -> Result<KmerMinHash, Error> {
-        let scaled = scaled_for_max_hash(max_hash);
-
-        let mut new_mh = KmerMinHash::new(
-            scaled,
-            self.ksize,
-            self.hash_function.clone(),
-            self.seed,
-            self.abunds.is_some(),
-            self.num,
-        );
-        if self.abunds.is_some() {
-            new_mh.add_many_with_abund(&self.to_vec_abunds())?;
+    pub fn downsample_max_hash(self, max_hash: u64) -> Result<KmerMinHash, Error> {
+        if self.max_hash == 0 {
+            // CTB: this is a num minhash. Should we just blithely return?
+            Ok(self)
         } else {
-            new_mh.add_many(&self.mins)?;
+            let scaled = scaled_for_max_hash(max_hash);
+            self.downsample_scaled(scaled)
         }
-        Ok(new_mh)
     }
 
     pub fn sum_abunds(&self) -> u64 {
@@ -778,13 +770,31 @@ impl KmerMinHash {
     // this could be improved by generating an HLL estimate while sketching instead
     // (for scaled minhashes)
     pub fn n_unique_kmers(&self) -> u64 {
-        self.size() as u64 * self.scaled() // + (self.ksize - 1) for bp estimation
+        self.size() as u64 * self.scaled() as u64 // + (self.ksize - 1) for bp estimation
     }
 
     // create a downsampled copy of self
-    pub fn downsample_scaled(&self, scaled: u64) -> Result<KmerMinHash, Error> {
-        let max_hash = max_hash_for_scaled(scaled);
-        self.downsample_max_hash(max_hash)
+    pub fn downsample_scaled(self, scaled: ScaledType) -> Result<KmerMinHash, Error> {
+        if self.scaled() == scaled || self.scaled() == 0 {
+            Ok(self)
+        } else if self.scaled() > scaled {
+            Err(Error::CannotUpsampleScaled)
+        } else {
+            let mut new_mh = KmerMinHash::new(
+                scaled,
+                self.ksize,
+                self.hash_function.clone(),
+                self.seed,
+                self.abunds.is_some(),
+                self.num,
+            );
+            if self.abunds.is_some() {
+                new_mh.add_many_with_abund(&self.to_vec_abunds())?;
+            } else {
+                new_mh.add_many(&self.mins)?;
+            }
+            Ok(new_mh)
+        }
     }
 
     pub fn inflate(&mut self, abunds_from: &KmerMinHash) -> Result<(), Error> {
@@ -996,7 +1006,7 @@ impl Clone for KmerMinHashBTree {
             mins: self.mins.clone(),
             abunds: self.abunds.clone(),
             current_max: self.current_max,
-            md5sum: Mutex::new(Some(self.md5sum())),
+            md5sum: Mutex::new(self.md5sum.lock().unwrap().clone()),
         }
     }
 }
@@ -1105,7 +1115,7 @@ impl<'de> Deserialize<'de> for KmerMinHashBTree {
 
 impl KmerMinHashBTree {
     pub fn new(
-        scaled: u64,
+        scaled: ScaledType,
         ksize: u32,
         hash_function: HashFunctions,
         seed: u64,
@@ -1147,7 +1157,7 @@ impl KmerMinHashBTree {
         self.max_hash
     }
 
-    pub fn scaled(&self) -> u64 {
+    pub fn scaled(&self) -> ScaledType {
         scaled_for_max_hash(self.max_hash)
     }
 
@@ -1356,13 +1366,14 @@ impl KmerMinHashBTree {
     }
 
     pub fn count_common(&self, other: &KmerMinHashBTree, downsample: bool) -> Result<u64, Error> {
-        if downsample && self.max_hash != other.max_hash {
-            let (first, second) = if self.max_hash < other.max_hash {
+        if downsample && self.scaled() != other.scaled() {
+            // downsample to the larger of the two scaled values
+            let (first, second) = if self.scaled() > other.scaled() {
                 (self, other)
             } else {
                 (other, self)
             };
-            let downsampled_mh = second.downsample_max_hash(first.max_hash)?;
+            let downsampled_mh = second.clone().downsample_scaled(first.scaled())?;
             first.count_common(&downsampled_mh, false)
         } else {
             self.check_compatible(other)?;
@@ -1487,13 +1498,14 @@ impl KmerMinHashBTree {
         ignore_abundance: bool,
         downsample: bool,
     ) -> Result<f64, Error> {
-        if downsample && self.max_hash != other.max_hash {
-            let (first, second) = if self.max_hash < other.max_hash {
+        if downsample && self.scaled() != other.scaled() {
+            // downsample to larger of two scaled
+            let (first, second) = if self.scaled() > other.scaled() {
                 (self, other)
             } else {
                 (other, self)
             };
-            let downsampled_mh = second.downsample_max_hash(first.max_hash)?;
+            let downsampled_mh = second.clone().downsample_scaled(first.scaled())?;
             first.similarity(&downsampled_mh, ignore_abundance, false)
         } else if ignore_abundance || self.abunds.is_none() || other.abunds.is_none() {
             self.jaccard(other)
@@ -1529,29 +1541,38 @@ impl KmerMinHashBTree {
     }
 
     // create a downsampled copy of self
-    pub fn downsample_max_hash(&self, max_hash: u64) -> Result<KmerMinHashBTree, Error> {
-        let scaled = scaled_for_max_hash(max_hash);
-
-        let mut new_mh = KmerMinHashBTree::new(
-            scaled,
-            self.ksize,
-            self.hash_function.clone(),
-            self.seed,
-            self.abunds.is_some(),
-            self.num,
-        );
-        if self.abunds.is_some() {
-            new_mh.add_many_with_abund(&self.to_vec_abunds())?;
+    pub fn downsample_max_hash(self, max_hash: u64) -> Result<KmerMinHashBTree, Error> {
+        if self.max_hash == 0 {
+            // CTB: this is a num minhash. Just blithely return.
+            Ok(self)
         } else {
-            new_mh.add_many(&self.mins())?;
+            let scaled = scaled_for_max_hash(max_hash);
+            self.downsample_scaled(scaled)
         }
-        Ok(new_mh)
     }
 
     // create a downsampled copy of self
-    pub fn downsample_scaled(&self, scaled: u64) -> Result<KmerMinHashBTree, Error> {
-        let max_hash = max_hash_for_scaled(scaled);
-        self.downsample_max_hash(max_hash)
+    pub fn downsample_scaled(self, scaled: ScaledType) -> Result<KmerMinHashBTree, Error> {
+        if self.scaled() == scaled || self.scaled() == 0 {
+            Ok(self)
+        } else if self.scaled() > scaled {
+            Err(Error::CannotUpsampleScaled)
+        } else {
+            let mut new_mh = KmerMinHashBTree::new(
+                scaled,
+                self.ksize,
+                self.hash_function.clone(),
+                self.seed,
+                self.abunds.is_some(),
+                self.num,
+            );
+            if self.abunds.is_some() {
+                new_mh.add_many_with_abund(&self.to_vec_abunds())?;
+            } else {
+                new_mh.add_many(&self.mins())?;
+            }
+            Ok(new_mh)
+        }
     }
 
     pub fn to_vec_abunds(&self) -> Vec<(u64, u64)> {
@@ -1646,6 +1667,8 @@ impl From<KmerMinHashBTree> for KmerMinHash {
         new_mh.mins = mins;
         new_mh.abunds = abunds;
 
+        new_mh.md5sum = other.md5sum;
+
         new_mh
     }
 }
@@ -1670,6 +1693,8 @@ impl From<&KmerMinHashBTree> for KmerMinHash {
         new_mh.mins = mins;
         new_mh.abunds = abunds;
 
+        new_mh.md5sum = Mutex::new(other.md5sum.lock().unwrap().clone());
+
         new_mh
     }
 }
@@ -1692,6 +1717,8 @@ impl From<KmerMinHash> for KmerMinHashBTree {
 
         new_mh.mins = mins;
         new_mh.abunds = abunds;
+
+        new_mh.md5sum = other.md5sum;
 
         new_mh
     }